requirements.txt
boto3==1.17.97
botocore==1.20.97
certifi==2021.5.30
chardet==4.0.0
idna==2.10
jmespath==0.10.0
python-dateutil==2.8.1
requests==2.25.1
requests-toolbelt==0.9.1
s3transfer==0.4.2
six==1.16.0
smartsheet-python-sdk==2.105.1
urllib3==1.26.5
